+++ float/src/lib.rs
@@ -81,6 +81,7 @@ pub mod round;
 mod round_ops;
 mod shift;
 mod sign;
+mod third_party;
 mod utils;
 
 pub use fbig::FBig;