+++ float/src/mul.rs
@@ -84,9 +84,36 @@ macro_rules! impl_add_sub_primitive_with_fbig {
 impl_add_sub_primitive_with_fbig!(u8 u16 u32 u64 u128 usize UBig i8 i16 i32 i64 i128 isize IBig);
 
 impl<R: Round, const B: Word> FBig<R, B> {
+    /// Compute the square of this number (`self * self`)
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use dashu_base::ParseError;
+    /// # use dashu_float::DBig;
+    /// let a = DBig::from_str_native("-1.234")?;
+    /// assert_eq!(a.square(), DBig::from_str_native("1.523")?);
+    /// # Ok::<(), ParseError>(())
+    /// ```
     #[inline]
     pub fn square(&self) -> Self {
         self.context.square(&self.repr).value()
+    }
+
+    /// Compute the square of this number (`self * self`)
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use dashu_base::ParseError;
+    /// # use dashu_float::DBig;
+    /// let a = DBig::from_str_native("-1.234")?;
+    /// assert_eq!(a.cubic(), DBig::from_str_native("-1.879")?);
+    /// # Ok::<(), ParseError>(())
+    /// ```
+    #[inline]
+    pub fn cubic(&self) -> Self {
+        self.context.cubic(&self.repr).value()
     }
 }
 
@@ -178,6 +205,44 @@ impl<R: Round> Context<R> {
         };
 
         let repr = Repr::new(f_repr.significand.square().into(), 2 * f_repr.exponent);
+        self.repr_round(repr).map(|v| FBig::new(v, *self))
+    }
+
+    /// Calculate the square of the floating point number under this context.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use dashu_base::ParseError;
+    /// # use dashu_float::DBig;
+    /// use dashu_base::Approximation::*;
+    /// use dashu_float::{Context, round::{mode::HalfAway, Rounding::*}};
+    ///
+    /// let context = Context::<HalfAway>::new(2);
+    /// let a = DBig::from_str_native("-1.234")?;
+    /// assert_eq!(context.cubic(&a.repr()), Inexact(DBig::from_str_native("-1.9")?, SubOne));
+    /// # Ok::<(), ParseError>(())
+    /// ```
+    pub fn cubic<const B: Word>(&self, f: &Repr<B>) -> Rounded<FBig<R, B>> {
+        check_inf(f);
+
+        // shrink the input operands if necessary
+        let max_precision = if self.is_limited() {
+            self.precision * 3
+        } else {
+            usize::MAX
+        };
+
+        let f_shrink;
+        let f_repr = if f.digits() > max_precision {
+            f_shrink = Context::<R>::new(max_precision).repr_round_ref(f).value();
+            &f_shrink
+        } else {
+            f
+        };
+
+        // TODO(next): increase dependency version on dashu_int because we use the new function cubic()
+        let repr = Repr::new(f_repr.significand.cubic(), 3 * f_repr.exponent);
         self.repr_round(repr).map(|v| FBig::new(v, *self))
     }
 }