+++ base/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.1 (WIP)
+
+- Add trait `Inverse`.
+- Implement `AbsCmp` and `AbsEq` for primitive types.
+
 ## 0.3.0
 
 ### Add